+++ examples/resources/routeros_system_certificate_scep_server/import.sh
@@ -1,3 +1,3 @@
 # The ID can be found via API or the terminal
 # The command for the terminal is -> /certificate/scep-server/print show-ids
-terraform import routeros_certificate_scep_server.example_scep_server "*1"
+terraform import routeros_system_certificate_scep_server.example_scep_server "*1"