+++ examples/resources/routeros_system_clock/resource.tf
@@ -1,0 +1,5 @@
+resource "routeros_system_clock" "set" {
+	date           = "2024-05-15"
+	time           = "17:58:11"
+	time_zone_name = "EST"
+}