+++ examples/resources/routeros_tool_bandwidth_server/resource.tf
@@ -1,0 +1,6 @@
+resource "routeros_tool_bandwidth_test_server" "test" {
+  enabled = true
+  authenticate = false
+  max_sessions = 100
+  allocate_udp_ports_from = 2000
+}