+++ routeros/resource_capsman_manager.go
@@ -37,11 +37,7 @@ func ResourceCapsManManager() *schema.Resource {
 			Default:     "none",
 			Description: "Device certificate.",
 		},
-		"enabled": {
-			Type:        schema.TypeBool,
-			Optional:    true,
-			Description: "Disable or enable CAPsMAN functionality.",
-		},
+		KeyEnabled: PropEnabled("Disable or enable CAPsMAN functionality."),
 		"generated_ca_certificate": {
 			Type:        schema.TypeString,
 			Computed:    true,